+++ timeline/tests/test_views.py
@@ -10,6 +10,7 @@ from timeline.tests.test_models import (
     TimelineEventTestBase,
     TimelineEventMixin,
     TimelineEventStatusMixin,
+    TEST_EXTRA_DATA,
 )
 
 
@@ -57,7 +58,7 @@ class ViewTestBase(
             user=self.user,
             event_name='test_event',
             description='description',
-            extra_data={'test_key': 'test_val'},
+            extra_data=TEST_EXTRA_DATA,
         )
 
 
@@ -126,7 +127,7 @@ class TestSiteTimelineView(ViewTestBase):
             user=self.user,
             event_name='test_event',
             description='description',
-            extra_data={'test_key': 'test_val'},
+            extra_data=TEST_EXTRA_DATA,
         )
 
     def test_get(self):
@@ -151,7 +152,7 @@ class TestSiteObjectTimelineView(ViewTestBase):
             user=self.user,
             event_name='test_event',
             description='description',
-            extra_data={'test_key': 'test_val'},
+            extra_data=TEST_EXTRA_DATA,
         )
         self.obj_ref = self.event_site.add_object(
             obj=self.user, label='user', name=self.user.username
@@ -187,7 +188,7 @@ class TestAdminTimelineView(ViewTestBase):
             user=self.user,
             event_name='test_event',
             description='description',
-            extra_data={'test_key': 'test_val'},
+            extra_data=TEST_EXTRA_DATA,
         )
 
     def test_get(self):